feat: include code review changes

+++ src/routes/task/task.routes.ts
@@ -6,7 +6,7 @@ const router = Router()
 router.post('/api/tasks', async(req, res) => {
 	try {
 		const {text} = req.body
-		if(text === '') {
+		if( /^\s*$/.test(text)) {
 			return res.status(400).json({message: 'Text is empty!'})
 		}
 		const data = new Task({
@@ -16,7 +16,6 @@ router.post('/api/tasks', async(req, res) => {
 			taskStatus: 'todo',
 			userId: req.user._id
 		})
-		console.log(req.user)
 		const dataToSave = data.save()
 		res.status(200).json(dataToSave)
 	}
@@ -27,7 +26,11 @@ router.post('/api/tasks', async(req, res) => {
 router.get('/api/tasks', async(req, res) => {
 	try {
 		const data = await Task.find({userId: req.user._id})
-		res.json(data)
+		if(data) {
+			res.json(data)
+		} else {
+			res.status(404).send({message: 'Tasks do not exists!'})
+		}
 	}
 	catch(error){
 		res.status(500).json({message: error.message})
@@ -36,25 +39,38 @@ router.get('/api/tasks', async(req, res) => {
 router.delete('/api/tasks/:id', async(req, res) => {
 	try {
 		const {id} = req.params
-		await Task.findByIdAndDelete(id)
-		res.send()
+		const data = await Task.findById(req.params.id)
+		if(!data) {
+			res.status(404).send({message: 'Task does not exist!'})
+		}
+		if(data.userId === req.user._id) {
+			await Task.findByIdAndDelete(id)
+			res.send()
+		} else {
+			res.status(403).send({message: 'Task does not belong to user'})
+		}
 	}
 	catch (error) {
-		res.status(400).json({message: error.message})
+		res.status(405).json({message: error.message})
 	}
 })
 router.patch('/api/tasks/:id', async(req, res) => {
 	try {
 		const updatedData = req.body
 		const {id} = req.params
+		const data = await Task.findById(req.params.id)
+		if(!data) {
+			res.status(404).send({message: 'Task does not exist!'})
+		}
 		const taskDescription = await Task.findById(req.params.id)
 		if(taskDescription.description === updatedData.description) {
 			return res.status(400).json({message: 'Description didn\'t change!'})
 		}
 		const result = await Task.findByIdAndUpdate(id, updatedData)
-		const data = await Task.findById(req.params.id)
 		if(data.userId === req.user._id) {
 			res.send(result)
+		} else {
+			res.status(403).send({message: 'Task dont belong to user'})
 		}
 	}
 	catch (error) {
@@ -66,10 +82,12 @@ router.get('/api/tasks/:id', async (req, res) => {
 		const data = await Task.findById(req.params.id)
 		if(data.userId === req.user._id) {
 			res.json(data)
+		}else {
+			res.status(403).send({message: 'Task dont belong to user'})
 		}
 	}
 	catch(error){
-		res.status(500).json({message: error.message})
+		res.status(405).json({message: error.message})
 	}
 })
